requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-environ==0.11.2
Pillow==10.1.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
crispy-bootstrap5>=2.0.1
django-parler==2.3
celery==5.3.4
redis==5.0.1
paypalrestsdk==1.13.1
django-templated-email==3.0.1
drf-spectacular==0.27.0
django-filter==23.5
easy-thumbnails==2.8.5
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3
